+++ Flow/src/utopia/flow/collection/immutable/TreeLike.scala
@@ -2,6 +2,7 @@ package utopia.flow.collection.immutable
 
 import utopia.flow.collection.template
 import utopia.flow.collection.template.TreeLike.AnyTree
+import utopia.flow.collection.CollectionExtensions._
 import utopia.flow.operator.EqualsExtensions.ApproxEquals
 
 /**
@@ -33,17 +34,28 @@ trait TreeLike[A, Repr <: TreeLike[A, Repr]] extends template.TreeLike[A, Repr]
 	// OTHER    ----------------
 	
 	/**
-	  * Creates a copy of this tree that contains a new child node
-	  * @param tree A child node to add
-	  * @return a copy of this tree with the provided child tree added
+	  * Creates a copy of this tree that contains the specified tree as one of its child nodes.
+	  * If this node already contains a matching child, merges it with this new node so that the children of both
+	  * are kept.
+	  * @param tree A child node to insert
+	  * @return a copy of this tree with the specified child included / merged in
 	  */
-	def +(tree: Repr) = createCopy(children = children :+ tree)
+	def :+(tree: Repr) =
+		createCopy(children = children.mergeOrAppend(tree) { _.nav ~== tree.nav } { _ :++ _.children })
 	/**
-	  * Creates a new tree that contains a child node with specified content
-	  * @param nodeContent Child node content
-	  * @return A copy of this tree with child node added
+	  * Creates a new tree that contains a child node with specified content.
+	  * If this node already contains such a child, doesn't add a new one.
+	  * @param nodeContent Child node content / nav
+	  * @return A copy of this tree including a child with that content / nav
 	  */
-	def +(nodeContent: A): Repr = this + newNode(nodeContent)
+	def :+(nodeContent: A): Repr = {
+		// Case: This node already contains that child => ignores
+		if (children.exists { _.nav ~== nodeContent })
+			repr
+		// Case: New child => inserts
+		else
+			createCopy(children = children :+ newNode(nodeContent))
+	}
 	
 	/**
 	  * Creates a new copy of this tree where the specified tree doesn't occur anywhere.
@@ -61,6 +73,64 @@ trait TreeLike[A, Repr <: TreeLike[A, Repr]] extends template.TreeLike[A, Repr]
 	def -(removedNav: A): Repr =
 		createCopy(children = children.filterNot { _.nav ~== removedNav }.map { _ - removedNav })
 	
+	/**
+	  * Merges these two trees together, if possible
+	  * @param other Another tree
+	  * @return Either:
+	  *         Left: This tree and the other tree, if there is no common root
+	  *         Right: A combination of these trees where all the children of the specified tree are included
+	  *         in this tree. Only returned if these trees have equal root nav.
+	  */
+	def +(other: Repr) = {
+		// Case: Identical root => merges children
+		if (nav ~== other.nav)
+			Right(this :++ other.children)
+		// Case: Different roots => returns as separate trees
+		else
+			Left(Pair(repr, other))
+	}
+	/**
+	  * Merges this tree with another set of trees, where possible.
+	  * The roots that have equal nav elements are merged together.
+	  * @param other Other trees
+	  * @return Merged trees. One entry for each unique root nav element.
+	  */
+	def ++(other: IterableOnce[Repr]) = {
+		// Combines the trees by nav, where possible (one tree at a time)
+		other.iterator.foldLeft(Vector(repr)) { (existingTrees, newTree) =>
+			existingTrees.mergeOrAppend(newTree) { _.nav ~== newTree.nav } { _ :++ _.children }
+		}
+	}
+	/**
+	  * Adds new children to this node using merging.
+	  * If there already exist children that have equal nav elements, they are combined with the newly introduced
+	  * children, yielding a single node per unique nav element, containing the children from both versions, merged
+	  * @param newChildren New children to merge into this tree
+	  * @return A copy of this tree that includes the new child tree structures
+	  */
+	def :++(newChildren: IterableOnce[Repr]): Repr = {
+		val iter = newChildren.iterator
+		// Case: There are actually some children to add => merges them in
+		if (iter.hasNext)
+			// Merges one new child at once
+			createCopy(children = iter.foldLeft(children) { (existingTrees, newTree) =>
+				existingTrees.mergeOrAppend(newTree) { _.nav ~== newTree.nav } { _ :++ _.children }
+			})
+		// Case: No children to add => No operation
+		else
+			repr
+	}
+	
+	/**
+	  * @param child A new child node
+	  * @return A copy of this tree with that child node added
+	  */
+	def withChildAdded(child: Repr) = createCopy(children = children :+ child)
+	/**
+	  * @param newChildren New child nodes
+	  * @return A copy of this tree with those child nodes added inder it
+	  */
+	def withChildrenAdded(newChildren: IterableOnce[Repr]) = createCopy(children = children ++ newChildren)
 	/**
 	  * Creates a new copy of this tree without the provided direct child node
 	  * @param child The child node that is removed from the direct children under this tree